+++ src/components/ListComments/ListComments.css
@@ -56,19 +56,6 @@
   margin-bottom: 0.5em;
 }
 
-.form-group textarea {
-  font-family: inherit;
-  width: 100%;
-  height: 10em;
-  border: 1px solid #ccc;
-  border-radius: 0.5em;
-  padding: 1em;
-  font-size: 1rem;
-  font-weight: 400;
-  color: #333;
-  resize: none;
-}
-
 .comment-submit {
   background-color: hsl(245, 58%, 51%);
   border: none;
