+++ src/components/NewComment/NewComment.css
@@ -1,0 +1,12 @@
+.comment-input {
+  font-family: inherit;
+  width: 100%;
+  height: 10em;
+  border: 1px solid #ccc;
+  border-radius: 0.5em;
+  padding: 1em;
+  font-size: 1rem;
+  font-weight: 400;
+  color: #333;
+  resize: none;
+}
